Ripples the statecache-lru main merge up. Resolved 4 conflicts by combining
the metrics-collector additions with the rippled cache/commitment changes:

- domain_shared.go: keep the kvmetrics MetricsCollectorProvider lookup +
  statecache-lru's new TrieConfig commitment-context ctor (cfg, branchCache).
- aggregator.go: keep the kvmetrics collector init + statecache-lru's
  oldestVisible; the MeteredGetLatestWithTxN/getLatestWithTxN methods now take
  *kvmetrics.DomainMetrics (the metrics relocation changeset -> kvmetrics).
- commitment_context.go: Metrics() returns *kvmetrics.DomainMetrics.
- eth_simulation.go: keep StartRequestMetrics(SourceRPC); the defer toggle is
  now sharedDomains.SetDeferCommitmentUpdates(false) (renamed in the refactor).

Validated: go build ./...; commitment + kvmetrics + execmodule reorg tests;
make lint clean.

+++ execution/cache/generic_cache.go
@@ -302,6 +302,6 @@ func (c *GenericCache[T]) PrintStatsAndReset(name string) {
 		"inserts", inserts, "evictions", evictions, "dropped", dropped,
 		"stale_evicted", staleEvicted, "epoch", c.epoch.Load(),
 		"entries", c.data.Len(), "size_mb", sizeBytes/(1024*1024),
-		"capacity_mb", c.capacityB/datasize.MB, "usage_pct", usagePct,
+		"capacity_mb", int64(c.capacityB/datasize.MB), "usage_pct", usagePct,
 	)
 }
